Adds browser tests:
- new grunt task `mocha-test` that will run tests using phantomjs
- new grunt task 'server` that will start local server up
  (you can then open http://localhost:8000/spec to run unit tests
  in your browser)
- new grunt task `watch` to watch file changes and run build again
- changed tokenizer tests: `match_tokens` and `be_token` are
  separate functions now to avoid `node` vs. `browser` vs. `phantomjs`
  shenanigans
