+++ spec/index.html
@@ -6,15 +6,16 @@
 </head>
 <body>
   <div id="mocha"></div>
-  <script src="/tmp/handlebars.browser.js"></script>
-  <script src="/vendor/mocha.js"></script>
-  <script src="/vendor/should.js"></script>
-  <script src="/tmp/tests/all-tests.js"></script>
-  <script>mocha.setup({ ui: 'bdd', timeout: 200 });</script>
+  <script src="../dist/handlebars.js"></script>
+  <script src="../vendor/mocha.js"></script>
+  <script src="../vendor/should.js"></script>
+  <script src="./env/common.js"></script>
+  <script src="./specHelpers/helpers.js"></script>
+  <script> mocha.setup({ ui: 'bdd', timeout: 200 }); </script>
+  <script src="../tmp/tests.js"></script>
   <script>
-    requireModule('spec/tests');
     if (window.mochaPhantomJS) { mochaPhantomJS.run(); }
-    else { mocha.run(); };
+    else { mocha.run(); }
   </script>
 </body>
 </html>
