+++ spec/tests/partials.js
